Rename auto.rs to sensors.rs
