+++ src/lib.rs
@@ -1,7 +1,7 @@
 #![allow(non_camel_case_types,non_snake_case,non_upper_case_globals)]
 
-pub mod auto;
+pub mod sensors;
 pub mod error;
 
-pub use auto::*;
+pub use sensors::*;
 pub use error::*;
